+++ examples/shopping_app/lib/features/checkout/models/checkout_shipping_request.g.dart
@@ -13,7 +13,7 @@
 
 part of 'checkout_shipping_request.dart';
 
-mixin _$CheckoutShippingRequest {
+mixin _$CheckoutShippingRequest implements Validatable {
   /// Validates this `CheckoutShippingRequest`.
   ///
   /// Usage: